\name{@TXDBOBJNAME@}
\docType{package}

\alias{@PKGNAME@-package}
\alias{@PKGNAME@}
\alias{@TXDBOBJNAME@}


\title{@PKGTITLE@}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at @PROVIDER@ on
  @RELEASEDATE@ and based on the @PROVIDERVERSION@
}

\author{@AUTHOR@}



\examples{
## load the library
##library(@PKGNAME@)
## list the contents that are loaded into memory
ls('package:@PKGNAME@')
## show the db object that is loaded by calling it's name
@PKGNAME@

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
