\name{@DBOBJNAME@}
\docType{package}

\alias{@PKGNAME@-package}
\alias{@PKGNAME@}
\alias{@DBOBJNAME@}


\title{@PKGTITLE@}

\description{
  This package contains a \code{\link[CompoundDb]{CompDb}} database
  providing compound annotations based on @SOURCE@ (version @SOURCEVERSION@).
  For examples and help on functions see the help pages from the
  \code{CompoundDb} package.
}

\author{@AUTHOR@}

\examples{
## load the library
library(@PKGNAME@)
## list the contents that are loaded into memory
ls('package:@PKGNAME@')
## show the db object that is loaded by calling it's name
@PKGNAME@

## for more examples see the CompoundDb package.


}

\keyword{package}
\keyword{data}
